Test iterators for both ProofMap variants [ECR-3871]:
Move tests of iterators to the base test to ensure they work
for both ProofMap variants.
